// src/enums/ENType.ts
/**
 * JS数据类型字符串枚举
 */
export enum ENType {
  STRING = 'string',
  NUMBER = 'number',
  BOOLEAN = 'boolean',
  UNDEFINED = 'undefined',
  OBJECT = 'object',
  ARRAY = 'array',
  FUNCTION = 'function',
  NULL = 'null',
  SYMBOL = 'symbol',
  DATE = 'date',
  REGEXP = 'regexp',
  ERROR = 'error',
  MAP = 'map',
  SET = 'set'
}